Clean up setup.sh
Remove setup-tools
Remove scss-lint from install steps

+++ setup.sh
@@ -1,26 +1,29 @@
 #!/bin/bash -e
 
 #####################################################
-#
 # Set up the Membership application
-#
 #####################################################
-
-echo " ____ ____ ____ ____ ____ ____ ____ ____ ____ ____ ";
-echo "||m |||e |||m |||b |||e |||r |||s |||h |||i |||p ||";
-echo "||__|||__|||__|||__|||__|||__|||__|||__|||__|||__||";
-echo "|/__\|/__\|/__\|/__\|/__\|/__\|/__\|/__\|/__\|/__\|";
-
-printf "\n\r"
-
-echo "  ^    ^    ^    ^    ^    ^    ^    ^  "
-echo " /f\  /r\  /o\  /n\  /t\  /e\  /n\  /d\ "
-echo "<___><___><___><___><___><___><___><___>"
 
 # fix the issues with bower installs
 git config --local url."https://".insteadOf git://
 
 cd frontend
+
+#####################################################
+# Bundle
+#####################################################
+
+printf "\n\r\n\r====================================\n\r\n\r"
+printf "> Installing Ruby gems..."
+printf "\n\r\n\r====================================\n\r\n\r"
+
+if hash bundler 2>/dev/null; then
+    bundle install
+else
+    printf "\nYou need to install bundler first:\n"
+    printf "\ngem install bundler. http://bundler.io/\n"
+    exit 1
+fi
 
 #####################################################
 # Install NPM modules
@@ -30,7 +33,13 @@ printf "\n\r\n\r====================================\n\r\n\r"
 printf "> Installing NPM modules ..."
 printf "\n\r\n\r====================================\n\r\n\r"
 
-npm install
+if hash grunt 2>/dev/null; then
+    npm install
+else
+    printf "\nYou need to install grunt-cli first:\n"
+    printf "\nnpm install -g grunt-cli\n"
+    exit 1
+fi
 
 #####################################################
 # Install Bower JS modules
@@ -40,9 +49,15 @@ printf "\n\r\n\r====================================\n\r\n\r"
 printf "> Installing bower JS modules in assets/javascripts ..."
 printf "\n\r\n\r====================================\n\r\n\r"
 
-pushd assets/javascripts
-rm -rf lib/bower-components
-bower install
+if hash bower 2>/dev/null; then
+    pushd assets/javascripts
+    rm -rf lib/bower-components
+    bower install
+else
+    printf "\nYou need to install bower first:\n"
+    printf "\nnpm install -g bower\n"
+    exit 1
+fi
 
 popd
